% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colNA.r
\name{colNA}
\alias{colNA}
\title{Number of NAs by column}
\usage{
colNA(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A numeric vector
}
\description{
Returns the number of NAs for each column of a data frame.
}
\examples{
set.seed(123)
(df <- data.frame(matrix(sample(c(NA, 1:10), 100, TRUE), ncol = 4)))
colNA(df)
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

