% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memUse.r
\name{memUse}
\alias{memUse}
\alias{print.memUse}
\title{Memory usage of saved objects}
\usage{
memUse(pos = 1, pattern)

\method{print}{memUse}(x, sort = c("size", "alphabetical"),
  decreasing = ifelse(sort == "size", TRUE, FALSE), n = 10, bytes = FALSE,
  ...)
}
\arguments{
\item{pos}{Which environment to use to list the saved objects (as
a position in the search list). See \code{\link{ls}}.}

\item{pattern}{An optional regular expression. See
\code{\link{ls}}.}

\item{x}{An object of class \code{memUse}.}

\item{sort}{Whether to sort the object list by \code{size} or
\code{alphabetical} order.}

\item{decreasing}{Logical.  Should the sort order be increasing or
decreasing?}

\item{n}{A single integer, giving the number of objects to
display.}

\item{bytes}{Logical. Whether to display the raw size in bytes.}
}
\value{
A data frame of classe \code{memUse} providing object
names, class and memory usage; or \code{0} if no object is present
in the specified environment.
}
\description{
Show memory usage of saved objects.
}
\section{Original URL}{

\url{http://stackoverflow.com/a/9839949}. See also the complete
Stack Overflow thread:
\url{https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

\examples{
## Create some data
rand <- rnorm(100)
data(iris)
foo <- function(x) print(x)
##'
## Check memory usage (and raw numbers in bytes)
memUse()
memUse()$bytes
}
\author{
Michael Hallquist, modified by Mathieu Basille
\email{basille@ufl.edu}
}
