% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsubset.r
\name{nsubset}
\alias{nsubset}
\title{Subset according to frequency of a column}
\usage{
nsubset(x, col, n, sign = c(">=", "<=", ">", "<", "=="), select,
  drop = FALSE)
}
\arguments{
\item{x}{object to be subsetted (e.g. \code{data.frame},
\code{matrix}, \code{Spatial*DataFrame}, etc.).}

\item{col}{name of the column which stores the grouping factor
(without quotes).}

\item{n}{the reference number for the frequency of the grouping
factor (must be a numeric of length 1).}

\item{sign}{any comparison sign (\code{>=} by default, \code{<=},
\code{>}, \code{<}, or \code{==}).}

\item{select}{expression, indicating columns to select (either
name(s) without quote or numeric indicating the column
number(s)).}

\item{drop}{passed on to '[' indexing operator.}
}
\value{
A subset of the object (with the same class).
}
\description{
Subset an object based on the frequency of a column (factor or
not).
}
\examples{
set.seed(1)
bla <- data.frame(value = rnorm(100), group = sample(letters[1:4],
    size = 100, replace = TRUE, prob = (1:4) * 10))
table(bla$group)
bli <- nsubset(bla, group, 25)
table(bli$group)
blo <- nsubset(bla, group, 25, sign = ">", select = 1, drop = TRUE)
head(blo)
length(blo)
}
\author{
Mathieu Basille \email{basille@ufl.edu}
}
