% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savepkglist.r
\name{savepkglist}
\alias{savepkglist}
\alias{attpkglist}
\alias{loadpkglist}
\alias{.loadpkglist}
\title{Load or save the list of attached packages}
\usage{
savepkglist(file = ".Rpackages")

attpkglist()

loadpkglist(file = ".Rpackages")

.loadpkglist()
}
\arguments{
\item{file}{The name of the file in which to save the list of
attached packages, or from which to load it. The path is relative
to the current working directory.}
}
\description{
Display, save or load the list of attached packages.
}
\details{
\code{attpkglist} simply lists all attached packages (i.e. not base
packages).

\code{savepkglist} saves the list of all attached packages in a file,
with one package per line.

\code{loadpkglist} loads a list of packages from a file. The file
should contain one package name per line, without quotes, and no empty
line. If the packages are not installed, the function sends a
warning.

\code{.loadpkglist} automatically loads the \code{.Rpackages} file at
startup (see the Note below).
}
\note{
To automatically load a \code{.Rpackages} list at startup,
add this in your \code{.Rprofile}:

\code{### Load packages at the start of R if the package list exists}

\code{basr:::.loadpkglist()}

Essentially, the function appends the list of packages at the end of
the \code{defaultPackages} option (see \code{\link[base]{options}} for
this option; see also \code{\link[base]{Startup}} for more details
about the initialization at start of an R session).
}
\examples{
\dontrun{savepkglist(file = "list.Rpackages")}
\dontrun{attpkglist()}
\dontrun{loadpkglist()}
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}
